requirements.txt
python-telegram-bot
gql
aiohttp
python-dotenv